Finally moved stuff around to flatten the directory hierarchy and tbblas is throwing exceptions instead of using asserts.

+++ capputils/reflection/ReflectableClass.h
@@ -272,7 +272,7 @@ protected: \
 #define DefineProperty(name, ...) \
   { \
     typedef TYPE_OF(name) Type; \
-    const unsigned Id = properties.size(); \
+    const size_t Id = properties.size(); \
     properties.push_back(new ::capputils::reflection::ClassProperty<Type>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, __VA_ARGS__, NULL)); \
     addressbook[#name] = (char*)&_##name - (char*)this; \
     CAPPUTILS_UNUSED(Id); \
@@ -281,7 +281,7 @@ protected: \
 #define ReflectableProperty(name, ...) \
 { \
   typedef TYPE_OF(name) Type; \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<Type>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, __VA_ARGS__, capputils::attributes::Reflectable<Type>(), NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
@@ -292,7 +292,7 @@ protected: \
 #define DefineProperty(name, arguments...) \
 { \
   typedef TYPE_OF(name) Type; \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<Type>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, ##arguments, NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
@@ -301,7 +301,7 @@ protected: \
 #define ReflectableProperty(name, arguments...) \
 { \
   typedef TYPE_OF(name) Type; \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<Type>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, ##arguments, capputils::attributes::Reflectable<Type>(), NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
@@ -337,7 +337,7 @@ protected: \
 
 #define DefineProperty(name, ...) \
 { \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<TYPE_OF(name)>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, __VA_ARGS__, NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
@@ -345,7 +345,7 @@ protected: \
 
 #define ReflectableProperty(name, ...) \
 { \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<TYPE_OF(name)>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, __VA_ARGS__, capputils::attributes::Reflectable<TYPE_OF(name)>(), NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
@@ -355,7 +355,7 @@ protected: \
 
 #define DefineProperty(name, arguments...) \
 { \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<TYPE_OF(name)>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, ##arguments, NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
@@ -363,7 +363,7 @@ protected: \
 
 #define ReflectableProperty(name, arguments...) \
 { \
-  const unsigned Id = properties.size(); \
+  const size_t Id = properties.size(); \
   properties.push_back(new ::capputils::reflection::ClassProperty<TYPE_OF(name)>(#name, _##name, ClassType ::get##name, ClassType ::set##name, ClassType ::reset##name, ##arguments, capputils::attributes::Reflectable<TYPE_OF(name)>(), NULL)); \
   addressbook[#name] = (char*)&_##name - (char*)this; \
   CAPPUTILS_UNUSED(Id); \
